Create README.md
In this project, you will be observing how different approaches result in different datasets based on product reviews and ratings in terms of positive and negative content (in reviews). So, this NLP application I have generated can help consumers to decide on products while comparing the review content with 3 different approaches.